+++ manual/release-notes.rst
@@ -18,6 +18,9 @@ more detail.
 
     - Avoid generating JSON with leading zeroes when converting real numbers.
 
+    - Detect and warn in check linearization when the xref stream reports the object containing a
+      compressed object to itself be a compressed object.
+
 12.4.0: August 9, 2026
   - Bug fixes